move tests in seperate folder
