+++ .vscode/settings.json
@@ -1,10 +1,12 @@
 {
     "clang-tidy.compilerArgs": [
         "-Iinc",
+        "-Itest",
         "--std=c17"
     ],
     "clang-tidy.compilerArgsBefore": [
         "-Iinc",
+        "-Itest",
         "--std=c17"
     ],
     "files.associations": {
