+++ makefile
@@ -1,11 +1,12 @@
 CBUILD = -Wall -Wextra --std=c17 -O0
-SRCS = main.c src/parser.c src/list.c
+SRCS = main.c test/tests.c src/parser.c src/list.c
+INC = -Iinc -Itest
 
 all: $(SRCS)
-	clang $^ $(CBUILD) -o main -fsanitize=address -DDEBUG -Iinc
+	clang $^ $(CBUILD) -o main -fsanitize=address -DDEBUG $(INC)
 
 main: $(SRCS)
-	clang $^ $(CBUILD) -o main -fsanitize=address -Iinc
+	clang $^ $(CBUILD) -o main -fsanitize=address -$(INC)
 
 clear:
 	@rm -f main
